fixes #211

+++ pyEX/studies/technicals/math.py
@@ -24,7 +24,7 @@ def acos(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.ACOS(df[col].values)
+    x = t.ACOS(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "acos": x})
 
 
@@ -43,7 +43,7 @@ def asin(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.ASIN(df[col].values)
+    x = t.ASIN(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "asin": x})
 
 
@@ -62,7 +62,7 @@ def atan(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.ATAN(df[col].values)
+    x = t.ATAN(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "atan": x})
 
 
@@ -81,7 +81,7 @@ def ceil(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.CEIL(df[col].values)
+    x = t.CEIL(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "ceil": x})
 
 
@@ -100,7 +100,7 @@ def cos(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.COS(df[col].values)
+    x = t.COS(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "cos": x})
 
 
@@ -119,7 +119,7 @@ def cosh(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.COSH(df[col].values)
+    x = t.COSH(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "cosh": x})
 
 
@@ -138,7 +138,7 @@ def exp(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.EXP(df[col].values)
+    x = t.EXP(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "exp": x})
 
 
@@ -157,7 +157,7 @@ def floor(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.FLOOR(df[col].values)
+    x = t.FLOOR(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "floor": x})
 
 
@@ -176,7 +176,7 @@ def ln(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.LN(df[col].values)
+    x = t.LN(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "ln": x})
 
 
@@ -195,7 +195,7 @@ def log10(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.LOG10(df[col].values)
+    x = t.LOG10(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "log10": x})
 
 
@@ -214,7 +214,7 @@ def sin(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.SIN(df[col].values)
+    x = t.SIN(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "sin": x})
 
 
@@ -233,7 +233,7 @@ def sinh(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.SINH(df[col].values)
+    x = t.SINH(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "sinh": x})
 
 
@@ -252,7 +252,7 @@ def sqrt(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.SQRT(df[col].values)
+    x = t.SQRT(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "sqrt": x})
 
 
@@ -271,7 +271,7 @@ def tan(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.TAN(df[col].values)
+    x = t.TAN(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "tan": x})
 
 
@@ -290,7 +290,7 @@ def tanh(client, symbol, timeframe="6m", col="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.TANH(df[col].values)
+    x = t.TANH(df[col].values.astype(float))
     return pd.DataFrame({col: df[col].values, "tanh": x})
 
 
@@ -310,7 +310,7 @@ def add(client, symbol, timeframe="6m", col1="open", col2="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.ADD(df[col1].values, df[col2].values)
+    x = t.ADD(df[col1].values.astype(float), df[col2].values.astype(float))
     return pd.DataFrame({col1: df[col1].values, col2: df[col2].values, "add": x})
 
 
@@ -330,7 +330,7 @@ def div(client, symbol, timeframe="6m", col1="open", col2="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.DIV(df[col1].values, df[col2].values)
+    x = t.DIV(df[col1].values.astype(float), df[col2].values.astype(float))
     return pd.DataFrame({col1: df[col1].values, col2: df[col2].values, "div": x})
 
 
@@ -350,7 +350,7 @@ def max(client, symbol, timeframe="6m", col="close", period=30):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    return t.MAX(df[col].values, period)
+    return t.MAX(df[col].values.astype(float), period)
 
 
 def maxindex(client, symbol, timeframe="6m", col="close", period=30):
@@ -369,7 +369,7 @@ def maxindex(client, symbol, timeframe="6m", col="close", period=30):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.MAXINDEX(df[col].values, period)
+    x = t.MAXINDEX(df[col].values.astype(float), period)
     return x, df[col].values[x]
 
 
@@ -389,7 +389,7 @@ def min(client, symbol, timeframe="6m", col="close", period=30):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    return t.MIN(df[col].values, period)
+    return t.MIN(df[col].values.astype(float), period)
 
 
 def minindex(client, symbol, timeframe="6m", col="close", period=30):
@@ -408,7 +408,7 @@ def minindex(client, symbol, timeframe="6m", col="close", period=30):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.MININDEX(df[col].values, period)
+    x = t.MININDEX(df[col].values.astype(float), period)
     return x, df[col].values[x]
 
 
@@ -428,7 +428,7 @@ def minmax(client, symbol, timeframe="6m", col="close", period=30):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    return t.MINMAX(df[col].values, period)
+    return t.MINMAX(df[col].values.astype(float), period)
 
 
 def minmaxindex(client, symbol, timeframe="6m", col="close", period=30):
@@ -447,7 +447,7 @@ def minmaxindex(client, symbol, timeframe="6m", col="close", period=30):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x, y = t.MINMAXINDEX(df[col].values, period)
+    x, y = t.MINMAXINDEX(df[col].values.astype(float), period)
     return x, df[col].values[x], y, df[col].values[y]
 
 
@@ -467,7 +467,7 @@ def mult(client, symbol, timeframe="6m", col1="open", col2="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.MULT(df[col1].values, df[col2].values)
+    x = t.MULT(df[col1].values.astype(float), df[col2].values.astype(float))
     return pd.DataFrame({col1: df[col1].values, col2: df[col2].values, "mult": x})
 
 
@@ -487,7 +487,7 @@ def sub(client, symbol, timeframe="6m", col1="open", col2="close"):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.SUB(df[col1].values, df[col2].values)
+    x = t.SUB(df[col1].values.astype(float), df[col2].values.astype(float))
     return pd.DataFrame({col1: df[col1].values, col2: df[col2].values, "sub": x})
 
 
@@ -507,5 +507,5 @@ def sum(client, symbol, timeframe="6m", col="close", period=30):
         DataFrame: result
     """
     df = client.chartDF(symbol, timeframe)
-    x = t.SUMMATION(df[col].values, period)
+    x = t.SUMMATION(df[col].values.astype(float), period)
     return pd.DataFrame({col: df[col].values, "sum": x})
